refactor for cleaner commit history as per recommendation
