+++ .gitignore
@@ -1,15 +1,15 @@
 # See https://docs.github.com/get-started/getting-started-with-git/ignoring-files for more about ignoring files.
 
 # Compiled output frontend
-/frontend/dist
-/frontend/tmp
-/frontend/out-tsc
-/frontend/bazel-out
+/dist
+/tmp
+/out-tsc
+/bazel-out
 
 # Node frontend
-/frontend/node_modules
-/frontend/npm-debug.log
-/frontend/yarn-error.log
+/node_modules
+/npm-debug.log
+/yarn-error.log
 
 # IDEs and editors
 .idea/
@@ -21,7 +21,7 @@
 *.sublime-workspace
 
 # Visual Studio Code
-.vscode/*
+.vscode
 !.vscode/settings.json
 !.vscode/tasks.json
 !.vscode/launch.json
@@ -29,13 +29,13 @@
 .history/*
 
 # Miscellaneous frontend
-/frontend/.angular/cache
-/frontend.sass-cache/
-/frontend/connect.lock
-/frontend/coverage
-/frontend/libpeerconnection.log
-/frontendtestem.log
-/frontend/typings
+/.angular/cache
+.sass-cache/
+/connect.lock
+/coverage
+/libpeerconnection.log
+testem.log
+/typings
 
 # System files
 .DS_Store
@@ -44,3 +44,4 @@ Thumbs.db
 # Backend
 /backend/.idea
 /backend/.venv
+node_modules
