+++ src/app/app.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+import { ButtonModule } from 'primeng/button';
+
+@Component({
+	selector: 'app-root',
+	standalone: true,
+	imports: [RouterOutlet, ButtonModule],
+	templateUrl: './app.component.html',
+	styleUrl: './app.component.scss',
+})
+export class AppComponent {
+	title = 'frontend';
+}
